+++ NINA.Plugin.TargetScheduler/Sequencer/TargetSchedulerCondition.cs
@@ -136,7 +136,7 @@ namespace NINA.Plugin.TargetScheduler.Sequencer {
 
         private bool HasRemainingTargets() {
             try {
-                Planner planner = new Planner(DateTime.Now, GetApplicableProfile(), GetProfilePreferences(), true);
+                Planner planner = new Planner(DateTime.Now, GetApplicableProfile(), GetProfilePreferences(), true, false);
                 bool result = planner.GetPlan(null) != null;
                 TSLogger.Info($"TargetSchedulerCondition check for remaining targets, continue={result}");
                 return result;
@@ -148,7 +148,7 @@ namespace NINA.Plugin.TargetScheduler.Sequencer {
 
         private bool HasActiveProjects() {
             try {
-                Planner planner = new Planner(DateTime.Now, GetApplicableProfile(), GetProfilePreferences(), true);
+                Planner planner = new Planner(DateTime.Now, GetApplicableProfile(), GetProfilePreferences(), true, false);
                 bool result = planner.HasActiveProjects(null);
                 TSLogger.Info($"TargetSchedulerCondition check for active projects, continue={result}");
                 return result;